explain,coord: move `explain_index`
Move `explain_index` to `inner/create_index.rs`.
